More vision progress
If you feel like selling your soul to Billy G, and going to the reitz
union's free connect 4 room - you can try it out! It finds all the
circles but has no way of finding the color or mapping it to an array.
Also the code is butt ugly and hideously innefficent watch out